+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 Changes:
 * Fix UI update error
+* Leave current connected channel when joining a second channel